test(ricer_test_tools)!: Move `fakes` module into test tool library
