--- tests/tools/mod.rs
@@ -1,6 +1,0 @@
-// SPDX-FileCopyrightText: 2024 Jason Pena <jasonpena@awkless.com>
-// SPDX-License-Identifier: GPL-2.0-or-later WITH GPL-CC-1.0
-
-pub mod stubs;
-
-pub mod fakes;
