+++ tests/tools/src/lib.rs
@@ -1,4 +1,7 @@
 // SPDX-FileCopyrightText: 2024 Jason Pena <jasonpena@awkless.com>
 // SPDX-License-Identifier: GPL-2.0-or-later WITH GPL-CC-1.0
 
+pub mod fakes;
+
 pub mod stubs;
+
